+++ lib/sums_up/result.rb
@@ -11,11 +11,20 @@ module SumsUp
   Result = SumsUp.define(failure: :error, success: :value) do
     # Yield, wrapping the result in Result.success, or wrap the raised error
     # in Result.failure.
-    def self.from_block
+    def self.from_block(error_class = StandardError)
       success(yield)
-    rescue StandardError => e
+    rescue error_class => e
       failure(e)
     end
+
+    def chain
+      match do |m|
+        m.success { |value| yield(value) }
+        m.failure self
+      end
+    end
+
+    alias_method(:flat_map, :chain)
 
     # Map a function across the successful value (if present).
     def map