+++ lib/sums_up/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## vX.Y.Z
+
+* Define `SumsUp::{Maybe,Result}#{chain,flat_map}`
+* `SumsUp::Result#from_block` now accepts an optional `Class` argument
+
 ## v1.1.0
 
 * Support rubies as low as 2.3